+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - An environment variable to enable/disable promotion codes.
 - An environment variable to change Payment Intent capture method.
 
+### Changed
+- Refactor crypto module `ctrKey` function.
+
 ### Removed
 - Properties quotes from shipping line item object.